fix: gracefully handle git blame errors
When running difflame on a commit that deletes an empty file, the git
blame command gets called when it should not. There's some logic in
get_blame_info_hunk() that could probably be adjusted to more cleanly
fix #6, but this try/catch statement does the trick, in a somewhat hacky
way.

Fixes: https://github.com/eantoranz/difflame/issues/6